requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.1
httpx==0.27.0
websockets==12.0
APScheduler==3.10.4
loguru==0.7.2